Finishing Auth

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama');
             $table->string('email')->unique();
             $table->string('alamat');
-            $table->integer('telpon')->unique();
+            $table->bigInteger('telpon')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
